task_3/requirements.txt
apache-airflow
google-cloud-storage
pandas
pyarrow